# api.yml
openapi: 3.0.3
info:
  title: Email Verification API
  version: 1.0.0
  description: |
    This API verifies the validity, risk, and deliverability of an email address.
servers:
  - url: https://microservices.com
paths:
  /emailVerify:
    post:
      summary: Verify and score an email address
      description: Verifies an email address for validity, risk, and deliverability, returning a detailed response object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationRequest"
      responses:
        "200":
          description: Email verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailVerificationAPIResponse"
        "400":
          description: Bad Request. The request was invalid or missing required parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
              example:
                success: false
                message: "Missing required parameter: email"
                errors: ["Parameter 'email' is required."]
        "500":
          description: Internal Server Error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
              example:
                success: false
                message: "A server error occurred. Please try again later."
                errors: ["Internal server error."]
        "503":
          description: Network error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
              example:
                success: false
                message: "A connection error occurred, Please check your connection again."
                errors: ["Netowork connection error."]
  /phoneValidate:
    post:
      summary: Verify and score an phone number
      description: Verifies an phone number for validity, risk.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneValidationRequest"
      responses:
        "200":
          description: Email verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneValidationResponse"
        "400":
          description: Bad Request. The request was invalid or missing required parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /proxyAndVpnDetection:
    post:
      summary: This is proxy and vpn detection api
      description: Verifies an ipAddress for validity, risk.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyAndVpnDetectionRequest"
      responses:
        "200":
          description: Proxy and vpn detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyAndVpnDetectionResponse"
        "400":
          description: Bad Request. The request was invalid or missing required parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    EmailVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          description: Email address to be verified.
    EmailVerificationResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: Indicates if the email address is valid (true/false).
        timedOut:
          type: boolean
          description: True if the email server did not respond within the expected time.
        disposable:
          type: boolean
          description: True if the email is from a temporary/disposable email provider.
        firstName:
          type: string
          description: First name associated with the email address, if available.
        deliverability:
          type: string
          enum: [high, medium, low]
          description: Likelihood that emails will be delivered to this address (e.g., high, medium, low).
        smtpScore:
          type: integer
          minimum: -1
          maximum: 3
          description: Numeric score showing how well the email server responds to SMTP checks.
        overallScore:
          type: integer
          minimum: 0
          maximum: 4
          description: General risk score for the email address, typically lower is better.
        catchAll:
          type: boolean
          description: True if the domain is a 'catch-all' and accepts all emails sent to it.
        generic:
          type: boolean
          description: True if the email is a generic address (e.g., info@, support@).
        common:
          type: boolean
          description: True if the email address is common or frequently used.
        dnsValid:
          type: boolean
          description: True if the domain has valid DNS records for email.
        honeypot:
          type: boolean
          description: True if the email is identified as a spam trap/honeypot.
        frequentComplainer:
          type: boolean
          description: True if this address is known to frequently mark emails as spam.
        suspect:
          type: boolean
          description: True if the email is suspected to be risky, fraudulent, or problematic.
        recentAbuse:
          type: boolean
          description: True if the email has recently been involved in abuse or spam activity.
        identityData:
          type: object
          description: Identity and enrichment information associated with the email address.
          properties:
            names:
              type: object
              description: An object containing names associated with the email address.
              properties:
                firstName:
                  type: string
                  description: Their first name.
                lastName:
                  type: string
                  description: Their last name.
                middleName:
                  type: string
                  description: Their middle name.
            addresses:
              type: object
              description: An object containing physical addresses associated with the email address.
              properties:
                address:
                  type: string
                  description: The street address.
                city:
                  type: string
                  description: The city.
                region:
                  type: string
                  description: The state or region.
                postalCode:
                  type: string
                  description: The postal code.
                country:
                  type: string
                  description: The country.
            emailAddresses:
              type: object
              description: An object containing other email addresses associated with the email address.
              properties:
                email:
                  type: string
                  description: The email address.
                firstSeen:
                  type: string
                  format: date-time
                  description: When the email address was first created.
                lastSeen:
                  type: string
                  format: date-time
                  description: When the email address was last used.
            phoneNumbers:
              type: object
              description: An object containing phone numbers associated with the email address.
              properties:
                formattedPhoneNumber:
                  type: string
                  description: The full phone number, including country code.
                country:
                  type: string
                  description: The country associated with the phone number.
            age:
              type: object
              description: An object containing ages of people associated with the email address.
              properties:
                human:
                  type: integer
                  description: How many years old the person is.
                birthdayIso:
                  type: string
                  format: date-time
                  description: Their date of birth in ISO format.
                birthdayTimestamp:
                  type: integer
                  description: Their birthday in timestamp format.
                birthdateCertainty:
                  type: boolean
                  description: Whether we are certain the birthdate is correct.
        fraudScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Numeric score (usually 0-100) representing the risk of fraud for this email.
        leaked:
          type: boolean
          description: True if this email address has appeared in data leaks or breaches.
        suggestedDomain:
          type: string
          description: Alternative domain suggestion if the original domain might be a typo.
        domainVelocity:
          type: string
          enum: [high, medium, low, none]
          description: Indicates how quickly the domain is being used/registered (e.g., high for many new accounts at once).
        domainTrust:
          type: string
          enum: [trusted, positive, neutral, suspicious, malicious, notRated]
          description: Level of trust for the domain (e.g., trusted, untrusted).
        userActivity:
          type: string
          enum: [high, medium, low, none]
          description: Level of past user activity associated with this email (high, medium, low).
        associatedNames:
          type: object
          description: Names associated with the email address (may require higher subscription).
          properties:
            status:
              type: string
              description: Status of data availability (e.g., "Enterprise Plus or higher required.").
            names:
              type: array
              items:
                type: string
              description: List of associated names.
        associatedPhoneNumbers:
          type: object
          description: Phone numbers associated with the email address (may require higher subscription).
          properties:
            status:
              type: string
              description: Status of data availability (e.g., "Enterprise Plus or higher required.").
            phoneNumbers:
              type: array
              items:
                type: string
              description: List of associated phone numbers.
        firstSeen:
          type: object
          description: When this email was first detected by the service.
          properties:
            human:
              type: string
              description: Human-readable representation of when first seen.
            timestamp:
              type: integer
              description: Unix timestamp of when first seen.
            iso:
              type: string
              format: date-time
              description: ISO 8601 date/time of when first seen.
        domainAge:
          type: object
          description: Data on when the domain was first registered or seen.
          properties:
            human:
              type: string
              description: Human-readable representation of domain age.
            timestamp:
              type: integer
              description: Unix timestamp of domain age.
            iso:
              type: string
              format: date-time
              description: ISO 8601 date/time of domain age.
        success:
          type: boolean
          description: True if the API request was completed successfully.
        spamTrapScore:
          type: string
          enum: [high, medium, low, none]
          description: Indicates spam trap risk (e.g., none, low, medium, high).
        riskyTld:
          type: boolean
          description: True if the top-level domain (TLD) is considered risky (e.g., .xyz, .top).
        spfRecord:
          type: boolean
          description: True if the domain has a valid SPF (Sender Policy Framework) record.
        dmarcRecord:
          type: boolean
          description: True if the domain has a valid DMARC record.
        sanitizedEmail:
          type: string
          description: The standardized/cleaned version of the input email address.
        mxRecords:
          type: array
          items:
            type: string
          description: List of mail exchange (MX) servers for the domain.
        requestId:
          type: string
          description: Unique identifier for this API request, useful for troubleshooting.
        aRecords:
          type: array
          items:
            type: string
          description: List of A (address) records for the domain, representing IP addresses.
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    EmailVerificationAPIResponse:
      type: object
      properties:
        apiResponse:
          $ref: "#/components/schemas/APIResponse"
        result:
          $ref: "#/components/schemas/EmailVerificationResponse"
    PhoneValidationRequest:
      type: object
      required:
        - phoneNumber
      properties:
        phoneNumber:
          type: string
          description: Phone number to be verified.
        country:
          type: array
          items:
            type: string
          description: (Optional) Country codes to validate against (e.g., ['US', 'UK']).
        strictness:
          type: integer
          enum: [0, 1]
          description: (Optional) Strictness level for validation (0 = normal, 1 = strictest).
    PhoneValidationResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message from the API.
        success:
          type: boolean
          description: Indicates if the request was successful.
        formatted:
          type: string
          description: Formatted phone number in international format.
        local_format:
          type: string
          description: Phone number in local format.
        valid:
          type: boolean
          description: Whether the phone number is valid.
        fraud_score:
          type: integer
          description: Risk score (0-100, higher means riskier).
        recent_abuse:
          type: boolean
          description: Whether there has been recent abuse from this number.
        VOIP:
          type: boolean
          description: Whether the phone number is a VOIP line.
        prepaid:
          type: boolean
          description: Whether the number is prepaid.
        risky:
          type: boolean
          description: Whether the phone number is considered risky.
        active:
          type: boolean
          description: Whether the phone number is currently active.
        name:
          type: string
          description: Name of the data provider.
        carrier:
          type: string
          description: Carrier name.
        line_type:
          type: string
          description: Type of phone line (e.g., mobile, landline, toll free).
        country:
          type: string
          description: Country code or name.
        region:
          type: string
          description: Region or state.
        city:
          type: string
          description: City.
        timezone:
          type: string
          description: Timezone of the number.
        zip_code:
          type: string
          description: Zip or postal code.
        accurate_country_code:
          type: boolean
          description: Indicates if the detected country code is accurate.
        dialing_code:
          type: integer
          description: International dialing code.
        leaked:
          type: boolean
          description: Whether the number has been leaked.
        spammer:
          type: boolean
          description: Whether the number is associated with spam.
        mnc:
          type: string
          description: Mobile Network Code.
        mcc:
          type: string
          description: Mobile Country Code.
        request_id:
          type: string
          description: Unique ID for the request.
    Error:
      type: object
      properties:
        errorCode:
          type: integer
          description: Numeric code representing the error type.
        errorMessage:
          type: string
          description: Description of the error that occurred.
    ProxyAndVpnDetectionRequest:
      type: object
      properties:
        ipAddress:
          type: string
          example: "8.8.8.0"
          description: ipAddress to be verified.
    ProxyAndVpnDetectionResponse:
      type: object
      properties:
        message:
          type: string
          example: "Success."
        success:
          type: boolean
          example: true
        fraudScore:
          type: integer
          example: 25
        proxy:
          type: boolean
          example: false
        ISP:
          type: string
          example: "Mediacom Cable"
        organization:
          type: string
          example: "Mediacom Cable"
        ASN:
          type: integer
          example: 30036
        host:
          type: string
          example: "192-0-2-110.client.mchsi.com"
        countryCode:
          type: string
          example: "US"
        city:
          type: string
          example: "Houston"
        region:
          type: string
          example: "Texas"
        isCrawler:
          type: boolean
          example: false
        latitude:
          type: number
          format: float
          example: 29.7079
        longitude:
          type: number
          format: float
          example: -95.401
        zipCode:
          type: string
          example: "77001"
        timezone:
          type: string
          example: "America Matamoros"
        vpn:
          type: boolean
          example: false
        tor:
          type: boolean
          example: false
        activeVpn:
          type: boolean
          example: false
        activeTor:
          type: boolean
          example: false
        recentAbuse:
          type: boolean
          example: true
        botStatus:
          type: boolean
          example: false
        mobile:
          type: boolean
          example: false