# src/Microservices/Controllers/DefaultApi.cs
/*
 * Email Verification API
 *
 * This API verifies the validity, risk, and deliverability of an email address. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Microservices.Attributes;
using Microservices.Models;

namespace Microservices.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// Verify and score an email address
        /// </summary>
        /// <remarks>Verifies an email address for validity, risk, and deliverability, returning a detailed response object.</remarks>
        /// <param name="emailVerificationRequest"></param>
        /// <response code="200">Email verification result</response>
        /// <response code="400">Bad Request. The request was invalid or missing required parameters.</response>
        /// <response code="500">Internal Server Error. An unexpected error occurred on the server.</response>
        /// <response code="503">Network error.</response>
        [HttpPost]
        [Route("/emailVerify")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("EmailVerifyPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(EmailVerificationAPIResponse), description: "Email verification result")]
        [SwaggerResponse(statusCode: 400, type: typeof(APIResponse), description: "Bad Request. The request was invalid or missing required parameters.")]
        [SwaggerResponse(statusCode: 500, type: typeof(APIResponse), description: "Internal Server Error. An unexpected error occurred on the server.")]
        [SwaggerResponse(statusCode: 503, type: typeof(APIResponse), description: "Network error.")]
        public abstract Task<IActionResult> EmailVerifyPost([FromBody]EmailVerificationRequest emailVerificationRequest);

        /// <summary>
        /// Verify and score an phone number
        /// </summary>
        /// <remarks>Verifies an phone number for validity, risk.</remarks>
        /// <param name="phoneValidationRequest"></param>
        /// <response code="200">Email verification result</response>
        /// <response code="400">Bad Request. The request was invalid or missing required parameters.</response>
        /// <response code="500">Internal Server Error. An unexpected error occurred on the server.</response>
        [HttpPost]
        [Route("/phoneValidate")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PhoneValidatePost")]
        [SwaggerResponse(statusCode: 200, type: typeof(PhoneValidationResponse), description: "Email verification result")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Bad Request. The request was invalid or missing required parameters.")]
        [SwaggerResponse(statusCode: 500, type: typeof(Error), description: "Internal Server Error. An unexpected error occurred on the server.")]
        public abstract Task<IActionResult> PhoneValidatePost([FromBody]PhoneValidationRequest phoneValidationRequest);

        /// <summary>
        /// This is proxy and vpn detection api
        /// </summary>
        /// <remarks>Verifies an ipAddress for validity, risk.</remarks>
        /// <param name="proxyAndVpnDetectionRequest"></param>
        /// <response code="200">Proxy and vpn detection result</response>
        /// <response code="400">Bad Request. The request was invalid or missing required parameters.</response>
        /// <response code="500">Internal Server Error. An unexpected error occurred on the server.</response>
        [HttpPost]
        [Route("/proxyAndVpnDetection")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("ProxyAndVpnDetectionPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(ProxyAndVpnDetectionResponse), description: "Proxy and vpn detection result")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Bad Request. The request was invalid or missing required parameters.")]
        [SwaggerResponse(statusCode: 500, type: typeof(Error), description: "Internal Server Error. An unexpected error occurred on the server.")]
        public abstract Task<IActionResult> ProxyAndVpnDetectionPost([FromBody]ProxyAndVpnDetectionRequest proxyAndVpnDetectionRequest);
    }
}